+++ src/ChildDeferred.cpp
@@ -12,7 +12,8 @@ ChildDeferred::ChildDeferred(const QVector<Deferred::Ptr>& parents, bool trackRe
 	: Deferred(), m_lock(QMutex::Recursive), m_resolvedCount(0), m_rejectedCount(0)
 {
 	setLogInvalidActionMessage(false);
-	setParents(parents, trackResults);
+	setParents(parents);
+	setTrackParentResults(trackResults);
 }
 
 ChildDeferred::Ptr ChildDeferred::create(Deferred::Ptr parent, bool trackResults)
@@ -27,6 +28,9 @@ ChildDeferred::Ptr ChildDeferred::create(const QVector<Deferred::Ptr>& parents, 
 
 ChildDeferred::~ChildDeferred()
 {
+	checkDestructionInSignalHandler();
+
+	QMutexLocker locker(&m_lock);
 	/* We disconnect all parent Deferreds to avoid that they trigger
 	 * onParentDestroyed() and onParentRejected() when m_parents is released
 	 * and they are still pending.
@@ -36,53 +40,101 @@ ChildDeferred::~ChildDeferred()
 		QObject::disconnect(parent.data(), 0, this, 0);
 }
 
-void ChildDeferred::setParent(Deferred::Ptr parent, bool trackResults)
+void ChildDeferred::setParent(Deferred::Ptr parent)
 {
-	setParents(QVector<Deferred::Ptr>{parent}, trackResults);
+	setParents(QVector<Deferred::Ptr>{parent});
 }
 
-void ChildDeferred::setParents(const QVector<Deferred::Ptr>& parents, bool trackResults)
+void ChildDeferred::setParents(const QVector<Deferred::Ptr>& parents)
 {
 	QMutexLocker locker(&m_lock);
 
 	for (Deferred::Ptr oldParent : const_cast<const QVector<Deferred::Ptr>&>(m_parents))
-		QObject::disconnect(oldParent.data(), 0 , this, 0);
+		QObject::disconnect(oldParent.data(), 0, this, 0);
 
 	for (Deferred::Ptr parent : parents)
 		QObject::connect(parent.data(), &QObject::destroyed, this, &ChildDeferred::onParentDestroyed);
 
-	if (trackResults)
+	m_parents = parents;
+
+	// If it is enabled, "restart" the result tracking with the new parents
+	if (m_trackParentResults)
+		setTrackParentResults(true);
+}
+
+void ChildDeferred::addParent(Deferred::Ptr parent)
+{
+	QMutexLocker locker(&m_lock);
+
+	QObject::connect(parent.data(), &QObject::destroyed, this, &ChildDeferred::onParentDestroyed, Qt::UniqueConnection);
+
+	m_parents.append(parent);
+
+	if (m_trackParentResults)
+		trackParentResult(parent.data());
+}
+
+void ChildDeferred::setTrackParentResults(bool trackParentResults)
+{
+	QMutexLocker locker(&m_lock);
+
+	m_trackParentResults = trackParentResults;
+	if (m_trackParentResults)
 	{
 		m_resolvedCount = m_rejectedCount = 0;
 
-		for (Deferred::Ptr parent : parents)
+		for (Deferred::Ptr parent : const_cast<const QVector<Deferred::Ptr>&>(m_parents))
+			trackParentResult(parent.data());
+	}
+	else
+	{
+		for (Deferred::Ptr parent : const_cast<const QVector<Deferred::Ptr>&>(m_parents))
 		{
-			switch(parent->state())
-			{
-			case Resolved:
-				QTimer::singleShot(0, this, [this, parent]() {
-					this->onParentResolved(parent->data());
-				});
-				break;
-			case Rejected:
-				QTimer::singleShot(0, this, [this, parent]() {
-					this->onParentRejected(parent->data());
-				});
-				break;
-			case Pending:
-			default:
-				QObject::connect(parent.data(), &Deferred::resolved, this, &ChildDeferred::onParentResolved);
-				QObject::connect(parent.data(), &Deferred::rejected, this, &ChildDeferred::onParentRejected);
-			}
+			QObject::disconnect(parent.data(), &Deferred::resolved, this, &ChildDeferred::onParentResolved);
+			QObject::disconnect(parent.data(), &Deferred::rejected, this, &ChildDeferred::onParentRejected);
+		}
+	}
+}
+
+void ChildDeferred::trackParentResult(Deferred* parent)
+{
+	switch(parent->state())
+	{
+	case Resolved:
+		QTimer::singleShot(0, this, [this, parent]() {
+			this->onParentResolved(parent->data());
+		});
+		break;
+	case Rejected:
+		QTimer::singleShot(0, this, [this, parent]() {
+			this->onParentRejected(parent->data());
+		});
+		break;
+	case Pending:
+	default:
+		QObject::connect(parent, &Deferred::resolved, this, &ChildDeferred::onParentResolved, Qt::UniqueConnection);
+		QObject::connect(parent, &Deferred::rejected, this, &ChildDeferred::onParentRejected, Qt::UniqueConnection);
+	}
+}
+
+void ChildDeferred::onParentDestroyed(QObject* parent)
+{
+	QMutexLocker locker(&m_lock);
+	qCritical("Parent deferred %s is destroyed while child %s is still holding a reference.", qUtf8Printable(pointerToQString(parent)), qUtf8Printable(pointerToQString(this)));
+	QObject::disconnect(parent, 0, this, 0);
+	auto deferredParent = static_cast<Deferred*>(parent);
+	QVector<int> removeIndices;
+	for (int i = m_parents.size()-1; i >= 0; --i)
+	{
+		if (m_parents.at(i) == deferredParent)
+		{
+			removeIndices.append(i);
 		}
 	}
 
-	m_parents = parents;
-}
-
-void ChildDeferred::onParentDestroyed(QObject* parent) const
-{
-	qCritical("Parent deferred %s is destroyed while child %s is still holding a reference.", qUtf8Printable(pointerToQString(parent)), qUtf8Printable(pointerToQString(this)));
+	for (int i : const_cast<const QVector<int>&>(removeIndices))
+		m_parents.removeAt(i);
+		
 }
 
 void ChildDeferred::onParentResolved(const QVariant& value)
@@ -92,6 +144,7 @@ void ChildDeferred::onParentResolved(const QVariant& value)
 	Q_EMIT parentResolved(value);
 	if (m_resolvedCount == m_parents.size())
 	{
+		// All parents have been resolved
 		QList<QVariant> results;
 		for (Deferred::Ptr parent : const_cast<const QVector<Deferred::Ptr>&>(m_parents))
 			results.append(parent->data());
@@ -106,6 +159,7 @@ void ChildDeferred::onParentRejected(const QVariant& reason)
 	Q_EMIT parentRejected(reason);
 	if (m_rejectedCount == m_parents.size())
 	{
+		// All parents have been rejected
 		QList<QVariant> reasons;
 		for (Deferred::Ptr parent : const_cast<const QVector<Deferred::Ptr>&>(m_parents))
 			reasons.append(parent->data());